requirements.txt
PyQt6
numpy
matplotlib
pyserial
